refactor(terminal): split commands into data + logic
